Rename excel to duckdb
