+++ dbt/include/excel/macros/adapters.sql
