+++ dbt/include/excel/macros/utils/any_value.sql
