+++ dbt/include/excel/macros/utils/splitpart.sql
